Making the AJAX Call
Let's head to our JavaScript! I'm going to write this a little bit different - though it would work either way: let's find all of the elements... there will be just one in this case... and loop over them with .each(). Indent the inner code, then close the extra function.

Now we can change the selector to this and... yea! We're basically doing the same thing as before. Inside the loop, fetch the URL with var autocompleteUrl = $ (this).data() to read that new attribute.

Finally, clear out the source attribute. Since we're using jQuery already, let's use it to make the AJAX call: $ .ajax() with a url option set to autocompleteUrl. That's it!

To handle the result, chain a .then() onto the Promise and pass a callback with a data argument. Let's see: our job is to execute the cb callback and pass it an array of the results.

Remember: in the controller, I'm returning all the user information on a users key. So, let's return data.users: that should return this entire array of data.

But also remember that, by default, the autocomplete library expects each result to have a value key that it uses. Obviously, our key is called email. To change that behavior, add displayKey: 'email'. I'll also add debounce: 500 - that will make sure that we don't make AJAX requests faster than once per half a second.

Ok... I think we're ready! Let's try this! Move back to your browser, refresh the page and clear out the author field... 'spac'... we got it! Though... it still returns all of the users - the geordi users should not match.

That's no surprise: our endpoint always returns every user.