Remember: in AccountController, we created an API endpoint that returns one User object. When we did that, we told the serializer to only serialize the groups called main. Look back in the User class. Ah, yes: we used the @groups() annotation to 'categorize' the fields we wanted into a group called main.
In AdminUtilityController, we can serialize that same group. Pass 200 as the second argument - this is the status code - we don't need any custom headers, but we do want to pass a groups option set to main... I know a lot of square brackets to do this.

Now go back and refresh. Got it! We could add a new serialization group to return even less - like maybe just the email. It's up to you.