Fixing Empty Value Case in the Data Transformer
Instead of failing, our data transformer return null when author box text is cleared.
Go back to EmailToUserTransformer. In reverseTransform(), if $ value is empty, just return. So, if the field is submitted empty, null should be passed to setAuthor().

But, hmm... the problem now is that, while it's technically ok to call setAuthor() with a null argument, we want that field to be required!

Re-submit the form! Oof - an integrity constraint violation: it's trying to save to the database with null set as the author_id column. We purposely made this required in the database and this is a great example of... messing up! We forget to add an important piece of business validation: to make the author required. No worries: open the Article class, find the $ author field and, above it, add @Assert\NotNull(). Give it a message: Please set an author.

Try that again. Excellent! This is the behavior - and error - we expect.

Next: how could we make our custom field type behave differently if it was used in different forms? Like, what if in one form, we want the user to be able to enter any user's email address but in another form we only want to allow the user to enter the email address of an admin user. Let's learn more about the power of form field options.