Our job is to use that to query for a User object and return it. And to do that, this class needs our UserRepository.
Time for some dependency injection! Add a constructor with UserRepository $ userRepository. I'll hit alt+enter and select 'Initialize Fields' to create that property and set it.

Normally... that's all we would need to do: we could instantly use that property below. But... this object is not instantiated by Symfony's container. So, we don't get our cool autowiring magic. Nope, in this case, we are creating this object ourselves! And so, we are responsible for passing it whatever it needs.

It's no big deal, but, we do have some more work.