This code requires a little bit of explanation. First, yea, the $ form->handleRequest() makes it look like the submitted data is being read and processed on every request, even the initial GET request that renders the form. But, that's not true! By default, handleRequest() only processes the data when this is a POST request. So, when the form is being submitted. When the form is originally loaded, handleRequest() sees that this is a GET request, does nothing, $ form->isSubmitted() returns false, and then the un-submitted form is rendered by Twig.
But, when we POST the form, ah, that's when handleRequest() does its magic. Because the form knows all of its fields, it grabs all of the submitted data from the $ request automatically and isSubmitted() returns true. Oh, and later, we'll talk about adding validation to our form. As you can guess, when validation fails, $ form->isValid() returns false.

So, wow! This controller does a lot, with very little code. And there are three possible flows.
1: if this is a GET request, isSubmitted() returns false and so the form is passed to Twig.
2: if this is a POST request but validation fails, isValid() returns false and so the form is again passed to Twig, but now it will render with errors. We'll see that later.
3: if this is a POST request and validation passes, both isSubmitted() and isValid() are true, and we finally get into the if block. ->getData() is how we access the final, normalized data that was submitted.

Phew! So, let's try it! Find your browser and create a very important article about the booming tourism industry on Mercury. Submit!

Yes! It dumps out exactly what we probably expected: an array with title and content keys. It's not too fancy yet, but it works nicely.