To make the flash message to look right we don't want to print the mb-5 class after the end of the nav element.
We could count the number of success flash messages, and if there are more than 0, do not print the mb-5 class. That's pretty simple, except for one huge problem! If we read the flash messages here to count them, that would also remove them! Our loop below would never do anything!
How can we work around that? By peeking at the flash messages.
The global app variable is actually an object called, conveniently, AppVariable! Press Shift+Shift and search for this so we can see exactly what it looks like.
Before, we used the getFlashes() method, which handles all the details of working with the Session object. But, if we need to 'peek', we need to work with the Session directly via the getSession() shortcut. It turns out, the 'flash messages' are stored on a sub-object called the 'flash bag'. This new longer code fetches the Session, gets that 'FlashBag' and calls peek() on it.