Setting the data_class Option
ConfigureOptions method is where you can set options that control how your form behaves. And, well... there aren't actually very many options. The most important, by far, is data_class. Set it to Article::class. This binds the form to that class. That little option changes everything.

Now, move back to your browser. Watch closely: right now both fields are simple text inputs... because we haven't configured them to be anything else. But, refresh!

Whoa! The content is now a textarea! We haven't talked about it yet, but we can, of course, configure how each field is rendered. By default, if you do nothing, everything renders as a text input. But, when you bind your form to a class, a special system - called the 'form type guessing' system - tries to guess the proper 'type' for each field. It notices that the  property on Article is a longer text Doctrine type. And so, it basically says:

  Hey peeps! This content field looks pretty big! So, let's use a textarea field type by default.

Anyways, form field type guessing is a cool feature. But, it is actually not the super important thing that just happened.

Create another breaking news story. Click Create and... yes! Check it out! $ form->getData() is now an Article object! And the title and content properties are already set! This is the power of the data_class option.